+++ handler/on_interaction_create.go
@@ -1,0 +1,14 @@
+package handler
+
+import (
+	"atomicnicos.me/digital-overdose-bot/extensions"
+
+	"github.com/bwmarrin/discordgo"
+)
+
+func OnInteractionCreate(s *discordgo.Session, i *discordgo.InteractionCreate) {
+	if h, ok := extensions.CommandHandlers[i.ApplicationCommandData().Name]; ok {
+		h(s, i)
+	}
+
+}