+++ main.go
@@ -7,10 +7,8 @@ Current features include:
 package main
 
 import (
-	"errors"
 	"flag"
 	"fmt"
-	"io"
 	"log"
 	"os"
 	"os/signal"
@@ -18,45 +16,19 @@ import (
 
 	"atomicnicos.me/digital-overdose-bot/common"
 	cron "atomicnicos.me/digital-overdose-bot/cron"
-	ext "atomicnicos.me/digital-overdose-bot/ext"
+	"atomicnicos.me/digital-overdose-bot/extensions"
+	"atomicnicos.me/digital-overdose-bot/handler"
 	"github.com/bwmarrin/discordgo"
 	"github.com/go-co-op/gocron"
 )
 
-var VERSION = "0.1.5-hotfix"
-
 // The Discord session, used for state management.
 var s *discordgo.Session
-var cronScheduler *gocron.Scheduler
 
 // Parses the flags provided in `argv`, then loads any overwrites from the .env file.
 func init() {
 	flag.Parse()
 	common.LoadEnvOrFlags()
-}
-
-// Initializes logging to a ./log/<datetime>.log file (if it can).
-func init() {
-	canWriteToFile := true
-	if _, err := os.Stat("./log/"); errors.Is(err, os.ErrNotExist) {
-		err := os.Mkdir("./log/", os.ModePerm)
-		if err != nil {
-			canWriteToFile = false
-			log.Println(err)
-		}
-	}
-
-	mw := io.MultiWriter()
-	if canWriteToFile {
-		f, err := os.OpenFile(fmt.Sprintf("log/%v-bot.log", time.Now().Format("2006-01-02-15-04-05")), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
-		if err != nil {
-			log.Fatalf("error opening file: %v", err)
-		}
-		mw = io.MultiWriter(os.Stdout, f)
-	} else {
-		mw = io.MultiWriter(os.Stdout)
-	}
-	log.SetOutput(mw)
 }
 
 // Starts up the bot.
@@ -71,23 +43,15 @@ func init() {
 	s.Identify.Intents = discordgo.IntentsAllWithoutPrivileged | discordgo.IntentsGuildMembers
 }
 
-var (
-	dmPermission                   = false
-	defaultMemberPermissions int64 = discordgo.PermissionManageServer
-	commands                       = ext.Commands
-	commandHandlers                = ext.CommandHandlers
-)
-
 func init() {
 	common.InitializeLogging()
 }
 
 func init() {
-	s.AddHandler(func(s *discordgo.Session, i *discordgo.InteractionCreate) {
-		if h, ok := ext.CommandHandlers[i.ApplicationCommandData().Name]; ok {
-			h(s, i)
-		}
-	})
+	s.AddHandler(handler.OnReady)
+	s.AddHandler(handler.OnInteractionCreate)
+	s.AddHandler(handler.OnMessage)
+	s.AddHandler(handler.OnJoin)
 }
 
 // Initializes cron engine, and subsequently registers scheduled functions
@@ -114,11 +78,6 @@ func init() {
 
 // Entry point, loads system and performs clean-up
 func main() {
-	s.AddHandler(func(s *discordgo.Session, r *discordgo.Ready) {
-		log.Printf("Running in version: v%v", VERSION)
-		log.Printf("Logged in as: '%v#%v'", s.State.User.Username, s.State.User.Discriminator)
-	})
-
 	err := s.Open()
 
 	if err != nil {
@@ -126,10 +85,10 @@ func main() {
 	}
 
 	log.Println("Adding commands...")
-	registeredCommands := make([]*discordgo.ApplicationCommand, len(ext.Commands))
+	registeredCommands := make([]*discordgo.ApplicationCommand, len(extensions.Commands))
 
 	// Registers all of the commands in the designated server.
-	for i, v := range ext.Commands {
+	for i, v := range extensions.Commands {
 		cmd, err := s.ApplicationCommandCreate(s.State.User.ID, *common.GuildID, v)
 		if err != nil {
 			log.Panicf("Cannot create '%v' command: %v", v.Name, err)
@@ -144,6 +103,15 @@ func main() {
 	signal.Notify(stop, os.Interrupt)
 	signal.Notify(stop, os.Kill)
 	log.Println("Press Ctrl+C to exit")
+
+	common.LogAndSend(fmt.Sprintf(":robot::rotating_light: is currently running on version `%v%v", common.VERSION, func() string {
+		if *common.CURRENTLY_DEV {
+			return " (DEV VERSION)`"
+		} else {
+			return "`"
+		}
+	}()), s)
+
 	<-stop
 
 	log.Println("")